requirements.txt
# Existing dependencies
streamlit
python-dotenv
Pillow
web3
eth-account
reportlab
PyPDF2
cryptography

# New dependencies for PDF digital signature
pyHanko>=0.12.0
pyhanko-certvalidator>=0.19.5
asn1crypto>=1.4.0
oscrypto>=1.3.0
certifi
click
pyyaml
tzlocal
uritools
qrcode[pil]

# Optional dependencies for better PDF support
pymupdf
pdfplumber